+++ VirtoCommerce.Storefront.Tests/Scriban/MyFunctions.cs
@@ -1,0 +1,11 @@
+namespace VirtoCommerce.Storefront.Tests.Scriban
+{
+    public class MyFunctions
+    {
+        public static string T(object input, params object[] variables)
+        {
+            return string.Format(input.ToString(), variables);
+        }
+
+    }
+}
