+++ VirtoCommerce.Storefront.Tests/Scriban/Tag.cs
@@ -1,0 +1,10 @@
+namespace VirtoCommerce.Storefront.Tests.Scriban
+{
+    public class Tag
+    {
+        public string GroupType { get; set; }
+        public string Lower { get; set; }
+        public string Upper { get; set; }
+
+    }
+}
