+++ VirtoCommerce.Storefront.Tests/Scriban/TestContext.cs
@@ -1,0 +1,10 @@
+using VirtoCommerce.Storefront.Model.Common;
+using VirtoCommerce.Storefront.Model.Catalog;
+
+namespace VirtoCommerce.Storefront.Tests.Scriban
+{
+    public class TestContext
+    {
+        public IMutablePagedList<Product> Products { get; set; }
+    }
+}
